test: add children tests and separate children arbitraries
